% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_psi.R
\name{default_psi}
\alias{default_psi}
\title{Default psi list}
\usage{
default_psi()
}
\value{
List with 4 different local features psi
}
\description{
List with the same local feature functions psi as in the original paper
}
\examples{
default_psi()

}
