% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwfGAM.R
\name{xwfGAM}
\alias{xwfGAM}
\title{Evaluate the GAM}
\usage{
xwfGAM(wL, wR, y, z = NULL)
}
\arguments{
\item{wL}{Matrix with left extrema-weighted features}

\item{wR}{Matrix with right extrema-weighted features}

\item{y}{Binary vector with outcomes}

\item{z}{Optional matrix z with extra, linear predictors}
}
\description{
Evaluate the generalized additive model for a set of computed extrema-weighted features
}
\examples{
xwf:::xwfGAM(wL = rep(1:45, 10), wR = rep(1:90, 5), y = c(rep(0:1, 225)))

}
